# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de BarberMusic&Spa",
        "description": "Documentación para la API utilizada en el sistema de BarberMusic&Spa. Aquí puedes probar los endpoints de autenticación.",
        "contact": {
            "email": "soporte@barbermusic.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/Scheduling_agendamientos/agendamientos": {
            "get": {
                "tags": [
                    "Agendamientos"
                ],
                "summary": "Obtener todos los agendamientos",
                "operationId": "7e06b010977ded84f4973dc6b319308b",
                "responses": {
                    "200": {
                        "description": "Lista de agendamientos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Agendamientos"
                ],
                "summary": "Crear un nuevo agendamiento",
                "operationId": "dcf1d68c43b8f02cc5d724b099bbae5f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cliente_usuario_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "personal_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "servicio_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "sucursal_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "fecha_hora_inicio": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-12-25 10:00:00"
                                    },
                                    "fecha_hora_fin": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-12-25 10:30:00"
                                    },
                                    "precio_final": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 250
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "PROGRAMADA"
                                    },
                                    "notas_cliente": {
                                        "type": "string",
                                        "example": "Por favor, ser puntual."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Agendamiento creado exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Agendamiento creado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/Scheduling_agendamientos/agendamientos/{id}": {
            "get": {
                "tags": [
                    "Agendamientos"
                ],
                "summary": "Obtener un agendamiento por ID",
                "operationId": "e6f3b1120d6928aa359af907fcd2775d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del agendamiento"
                    },
                    "404": {
                        "description": "Agendamiento no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Agendamientos"
                ],
                "summary": "Actualizar un agendamiento",
                "operationId": "adc34adcdc81643bf352545ec15f8346",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "estado": {
                                        "type": "string",
                                        "example": "CONFIRMADA"
                                    },
                                    "notas_internas": {
                                        "type": "string",
                                        "example": "Cliente contactado por teléfono."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Agendamiento actualizado"
                    },
                    "404": {
                        "description": "Agendamiento no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Agendamientos"
                ],
                "summary": "Eliminar un agendamiento (Soft Delete)",
                "operationId": "2080c63635838e5a3d465a4f225c9f92",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Agendamiento eliminado"
                    },
                    "404": {
                        "description": "Agendamiento no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Introduce el token de acceso (JWT o Sanctum) sin la palabra 'Bearer'.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}